Address bottlenecks in messaging channels
Do not block when output chan is full. Drop messages if output chan is unconsumed by caller.
InfiniteChan construct to extend a buffered channel with queue.
Do not block if await chan is full. Spin and try again until message is pushed or request is done and channel is removed.